User: Header: Home Link
Home link is a variable link which depends on the user linked.

As a substitute now Header.xml contains $$UNAME$$ to be replaced with
user name from useridentity.

This replacement is done as first step in __user.aspx.cs which home link
is clicked.

TODO: This have an error that link name contains $$UNAME$$ string. This
should be corrected when proper method is found.

Signed-off-by: NKK <nkk@gmail.com>