Registration: State Ajax calls
State DropDownList should vary as country value is varied.
This is done through AJAX request. If state list for some country is
empty then this is taken as 5000(a big value)

ToDo: State DropDown List values were changing based on Country name
by Client Script. This is resulting in page validation failure.

To save page validation failure Page's EnableEventValidation is set to
false. But this is not correct  solution and is security threat.

Their are solutions available for this. Comments are added in AddUser
function for these solution.

Signed-off-by: NKK <nkk@gmail.com>