Master: LeftRight: Login Control with cookies
LeftRight master adds one more login control(in the right panel). Now if
user cookie is set or session variable set then login control in right
should not be visible(note login control if user is logged in changes to
log out button but this is available in header). This change checks if
user logged in or not else make panel visibility to false.

Signed-off-by: NKK <nkk@gmail.com>