Base Class: Base Activation Page
Again in the heirarchy of BasePage classe added class for Base
Activation.

This class is used to verify Activation Links, e.g. email verification
link.

Take following example
1) same page email-verify.aspx is used to generate email verification
link and verify it
2) Which particular functionality to be displayed depends on Base
Activation class.

If on load query strings corresponding to Activation are present then it
alerts by setting Activation Variable that this is going to be
Activation other wise it is transparent and works as BaseLogged in
class.

On receiving Activation Link this class verifies proper activation link
and sets proper variable so that derived class can know if link was
verified or not.

Signed-off-by: NKK <nkk@gmail.com>