Base Logged in: Update for modal window
Current implementation is empty implementation. But proper structure is
added.

Ideally for pages which require to be logged in should be derived from
this class only. This flow is not working at preset as script manager
should be called first. But with class inheritence page load happens
late.

Need to see proper solution for this.

Signed-off-by: NKK <nkk@gmail.com>